menerapkan implementasi authorization

+++ src/api/notes/handler.js
@@ -16,7 +16,13 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+      const noteId = await this._service.addNote({
+        title,
+        body,
+        tags,
+        owner: credentialId,
+      });
       const response = h.response({
         status: 'success',
         message: 'Catatan berhasil ditambahkan',
@@ -47,8 +53,10 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
+
     return {
       status: 'success',
       data: {
@@ -60,7 +68,11 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
+
       return {
         status: 'success',
         data: {
@@ -77,7 +89,7 @@ class NotesHandler {
         return response;
       }
 
-      //Server Error!
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.',
@@ -92,6 +104,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
+
       await this._service.editNoteById(id, request.payload);
       return {
         status: 'success',
@@ -118,9 +133,12 @@ class NotesHandler {
     }
   }
 
-  async deleteNoteByIdHandler(request) {
+  async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
+
       await this._service.deleteNoteById(id);
       return {
         status: 'success',
